requirements.txt
# ============================================================
# Lung Disease Detection AI — Python Dependencies
# ============================================================

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Database ORM & Migrations
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Machine Learning / Deep Learning
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
scikit-learn==1.3.2

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.0.1
matplotlib==3.7.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.4
flask-cors==4.0.0
flask-talisman==1.1.0

# Development
gunicorn==21.2.0
